requirements.txt
torch>=1.9.0
torchvision>=0.13.0
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
scapy>=2.4.5
tqdm>=4.50.0
pyyaml>=5.4.0
matplotlib>=3.3.0
seaborn>=0.11.0
pytest>=6.0.0
black>=21.5b2
isort>=5.9.0
flwr>=0.19.0
ray[rllib]>=1.8.0